+++ src/UnitTestEx/Logging/LoggerBase.cs
@@ -34,7 +34,11 @@ namespace UnitTestEx.Logging
         public string Name { get; }
 
         /// <inheritdoc />
+#if NET7_0_OR_GREATER
+        public IDisposable BeginScope<TState>(TState state) where TState : notnull => _scopeProvider.Push(state);
+#else
         public IDisposable BeginScope<TState>(TState state) => _scopeProvider.Push(state);
+#endif
 
         /// <inheritdoc />
         public bool IsEnabled(LogLevel logLevel) => true;